+++ src/constants.rs
@@ -3,7 +3,7 @@ use std::time::Duration;
 use crate::funds::FundsAmount;
 
 pub const STARTING_FUNDS: FundsAmount = 5000;
-pub const NEW_MINION_PROGRESS: usize = 100;
+pub const NEW_MINION_PROGRESS: f64 = 100.0;
 
 pub const AUTOSAVE_INTERVAL: Duration = Duration::from_mins(5);
